ingestion-service/requirements.txt
google-cloud-aiplatform>=1.38.0
google-cloud-discoveryengine>=0.11.0
google-cloud-firestore>=2.14.0
google-cloud-storage>=2.14.0
msal>=1.26.0
requests>=2.31.0
beautifulsoup4>=4.12.0
markdownify>=0.11.6
pydantic>=2.5.0
fastapi>=0.100.0
uvicorn>=0.24.0